api/requirements.txt
Django>=4.0,<4.1.3
pytz>=2022
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8
django-cors-headers>=3.5.0